apps/api_v2/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3
pydantic>=2.11.2
pydantic-settings>=2.5.2
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
httpx>=0.28.1
pyjwt==2.9.0
cryptography>=45.0.0
clerk-backend-api==4.0.0
requests==2.32.3
aiofiles==24.1.0
svix==1.40.0
